+++ src/neoxp/Commands/BatchCommand.BatchFileCommands.cs
@@ -160,13 +160,28 @@ namespace NeoExpress.Commands
                 {
                     [Argument(0, Description = "Policy to set")]
                     [Required]
-                    internal PolicyName Policy { get; init; }
+                    internal PolicySettings Policy { get; init; }
 
                     [Argument(1, Description = "New Policy Value")]
                     [Required]
-                    internal long Value { get; set; }
+                    internal decimal Value { get; set; }
 
                     [Argument(2, Description = "Account to pay contract invocation GAS fee")]
+                    [Required]
+                    internal string Account { get; init; } = string.Empty;
+
+                    [Option(Description = "password to use for NEP-2/NEP-6 sender")]
+                    internal string Password { get; init; } = string.Empty;
+                }
+
+                [Command("sync")]
+                internal class Sync
+                {
+                    [Argument(0, Description = "Source of policy values. Must be path to policy settings JSON file")]
+                    [Required]
+                    internal string Source { get; } = string.Empty;
+
+                    [Argument(1, Description = "Account to pay contract invocation GAS fee")]
                     [Required]
                     internal string Account { get; init; } = string.Empty;