+++ src/neoxp/Commands/OracleCommand.Response.cs
@@ -50,7 +50,7 @@ namespace NeoExpress.Commands
                 }
                 catch (Exception ex)
                 {
-                    app.WriteException(ex);
+                    app.WriteException(ex, showInnerExceptions: true);
                     return 1;
                 }
             }