+++ src/neoxp/Commands/PolicyCommand.Get.cs
@@ -1,7 +1,7 @@
 using System;
-using System.ComponentModel.DataAnnotations;
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
+using Neo.Network.RPC;
 using NeoExpress.Models;
 
 namespace NeoExpress.Commands
@@ -18,29 +18,64 @@ namespace NeoExpress.Commands
                 this.chainManagerFactory = chainManagerFactory;
             }
 
-            [Argument(0, Description = "Policy to set")]
-            [Required]
-            internal PolicyName Policy { get; init; }
+            [Option(Description = "URL of Neo JSON-RPC Node\nSpecify MainNet (default), TestNet or JSON-RPC URL")]
+            internal string RpcUri { get; } = string.Empty;
 
             [Option(Description = "Path to neo-express data file")]
             internal string Input { get; init; } = string.Empty;
+
+            [Option(Description = "Output as JSON")]
+            internal bool Json { get; init; } = false;
 
             internal async Task<int> OnExecuteAsync(CommandLineApplication app, IConsole console)
             {
                 try
                 {
-                    var (chainManager, _) = chainManagerFactory.LoadChain(Input);
-                    using var expressNode = chainManager.GetExpressNode();
+                    var values = await GetPolicyValuesAsync().ConfigureAwait(false);
 
-                    var value = await expressNode.GetPolicyAsync(Policy).ConfigureAwait(false);
-                    await console.Out.WriteLineAsync($"{Policy}: {value}");
-
+                    if (Json)
+                    {
+                        await console.Out.WriteLineAsync(values.ToJson().ToString(true));
+                    }
+                    else
+                    {
+                        await WritePolicyValueAsync(console, $"             {nameof(PolicyValues.GasPerBlock)}", values.GasPerBlock);
+                        await WritePolicyValueAsync(console, $"    {nameof(PolicyValues.MinimumDeploymentFee)}", values.MinimumDeploymentFee);
+                        await WritePolicyValueAsync(console, $"{nameof(PolicyValues.CandidateRegistrationFee)}", values.CandidateRegistrationFee);
+                        await WritePolicyValueAsync(console, $"        {nameof(PolicyValues.OracleRequestFee)}", values.OracleRequestFee);
+                        await WritePolicyValueAsync(console, $"       {nameof(PolicyValues.NetworkFeePerByte)}", values.NetworkFeePerByte);
+                        await WritePolicyValueAsync(console, $"        {nameof(PolicyValues.StorageFeeFactor)}", values.StorageFeeFactor);
+                        await WritePolicyValueAsync(console, $"      {nameof(PolicyValues.ExecutionFeeFactor)}", values.ExecutionFeeFactor);
+                    }
                     return 0;
                 }
                 catch (Exception ex)
                 {
                     app.WriteException(ex);
                     return 1;
+                }
+            }
+
+            static Task WritePolicyValueAsync(IConsole console, string name, Neo.BigDecimal value)
+                => console.Out.WriteLineAsync($"{name}: {value.Value} ({value} GAS)");
+
+            static Task WritePolicyValueAsync(IConsole console, string name, uint value)
+                => console.Out.WriteLineAsync($"{name}: {value}");
+
+            async Task<PolicyValues> GetPolicyValuesAsync()
+            {
+                if (string.IsNullOrEmpty(RpcUri))
+                {
+                    var (chainManager, _) = chainManagerFactory.LoadChain(Input);
+                    using var expressNode = chainManager.GetExpressNode();
+                    return await expressNode.GetPolicyAsync().ConfigureAwait(false);
+                }
+                else 
+                {
+                    if (!TransactionExecutor.TryParseRpcUri(RpcUri, out var uri))
+                        throw new ArgumentException($"Invalid RpcUri value \"{RpcUri}\"");
+                    using var rpcClient = new RpcClient(uri);
+                    return await rpcClient.GetPolicyAsync().ConfigureAwait(false);
                 }
             }
         }