+++ src/neoxp/Extensions/Extensions.cs
@@ -21,18 +21,17 @@ namespace NeoExpress
     {
         public static void WriteException(this CommandLineApplication app, Exception exception, bool showInnerExceptions = false)
         {
-            var stackTraceOption = (CommandOption<bool>)app.GetOptions().Single(o => o.LongName == "stack-trace");
+            var showStackTrace = ((CommandOption<bool>)app.GetOptions().Single(o => o.LongName == "stack-trace")).ParsedValue;
 
-            var exceptionName = stackTraceOption.ParsedValue || showInnerExceptions 
-                ? $" [{exception.GetType()}]" : string.Empty;
-            app.Error.WriteLine($"\x1b[1m\x1b[31m\x1b[40m{exception.Message}{exceptionName}\x1b[0m");
-            if (stackTraceOption.ParsedValue) app.Error.WriteLine(exception.StackTrace);
+            app.Error.WriteLine($"\x1b[1m\x1b[31m\x1b[40m{exception.GetType()}: {exception.Message}\x1b[0m");
 
-            if (showInnerExceptions)
+            if (showStackTrace) app.Error.WriteLine($"\x1b[1m\x1b[37m\x1b[40m{exception.StackTrace}\x1b[0m");
+
+            if (showInnerExceptions || showStackTrace)
             {
                 while (exception.InnerException != null)
                 {
-                    app.Error.WriteLine($"  Inner Exception: {exception.InnerException.Message} [{exception.InnerException.GetType().Name}]");
+                    app.Error.WriteLine($"\x1b[1m\x1b[33m\x1b[40m\tInner {exception.InnerException.GetType().Name}: {exception.InnerException.Message}\x1b[0m");
                     exception = exception.InnerException;
                 }
             }
@@ -101,18 +100,9 @@ namespace NeoExpress
             return result;
         }
 
-        public static async Task<Neo.IO.Json.JObject> RpcSendAsync(this RpcClient @this, string method, params Neo.IO.Json.JObject[] paraArgs)
+        public static Task<PolicyValues> GetPolicyAsync(this RpcClient rpcClient)
         {
-            var request = new Neo.Network.RPC.Models.RpcRequest
-            {
-                Id = 1,
-                JsonRpc = "2.0",
-                Method = method,
-                Params = paraArgs
-            };
-
-            var response = await @this.SendAsync(request).ConfigureAwait(false);
-            return response.Result;
+            return ExpressNodeExtensions.GetPolicyAsync(script => rpcClient.InvokeScriptAsync(script));
         }
 
         public static bool TryGetSigningAccount(this ExpressChainManager chainManager, string name, string password, [MaybeNullWhen(false)] out Wallet wallet, [MaybeNullWhen(false)] out UInt160 accountHash)